backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic-settings
fredapi
fear-and-greed
finnhub-python
pandas
numpy
psycopg2-binary
asyncpg
sqlalchemy
google-genai
yfinance
redis
httpx
ta
beautifulsoup4
pytz
